made corrections to naming and saving a picture
